refactor(core): move configs, RunContext into core package
