+++ src/main/kotlin/com/jonaswanke/unicorn/action/Main.kt
@@ -2,7 +2,7 @@ package com.jonaswanke.unicorn.action
 
 import com.jonaswanke.unicorn.action.Action.throwError
 import com.jonaswanke.unicorn.api.*
-import com.jonaswanke.unicorn.commands.RunContext
+import com.jonaswanke.unicorn.core.RunContext
 import org.kohsuke.github.GHIssue
 import org.kohsuke.github.GHPullRequest
 import java.nio.file.FileSystems
