+++ src/main/kotlin/com/jonaswanke/unicorn/api/Files.kt
@@ -1,6 +1,6 @@
 package com.jonaswanke.unicorn.api
 
-import com.jonaswanke.unicorn.commands.RunContext
+import com.jonaswanke.unicorn.core.RunContext
 import net.swiftzer.semver.SemVer
 import java.io.BufferedWriter
 import java.io.File
