+++ src/main/kotlin/com/jonaswanke/unicorn/api/Git.kt
@@ -1,6 +1,6 @@
 package com.jonaswanke.unicorn.api
 
-import com.jonaswanke.unicorn.commands.RunContext
+import com.jonaswanke.unicorn.core.RunContext
 import com.jonaswanke.unicorn.utils.code
 import net.swiftzer.semver.SemVer
 import org.eclipse.jgit.api.GitCommand
