+++ src/main/kotlin/com/jonaswanke/unicorn/commands/BaseCommand.kt
@@ -7,6 +7,7 @@ import com.github.ajalt.clikt.parameters.options.default
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.types.file
 import com.jonaswanke.unicorn.console.ConsoleRunContext
+import com.jonaswanke.unicorn.core.RunContext
 import com.jonaswanke.unicorn.utils.TextIoConsoleWrapper
 import java.io.File
