+++ src/main/kotlin/com/jonaswanke/unicorn/commands/Commands.kt
@@ -1,6 +1,7 @@
 package com.jonaswanke.unicorn.commands
 
 import com.github.ajalt.clikt.core.subcommands
+import com.jonaswanke.unicorn.core.RunContext
 
 object Commands : BaseCommand() {
     init {
